+++ src/Diff2D/types.hpp
@@ -5,7 +5,7 @@
 
 typedef std::vector< array<real,1> >		coor_type;
 
-typedef std::vector< array<int,1> >		cell_count_type;
+typedef std::vector< array<size_t,1> >		cell_count_type;
 
 
 typedef multivec< 3, array<real,1> >				v_bou_vec_type;
